Refactoring of authContext
Heavy refactor of the authContext. Now user Data is persisted on page refresh.
Also, now saving the users roles so we have role based conditional rendering.
See the userDropdown comp for the syntax.

Moved the myTickets page back to where it should be. I moved it by accident.

mild refactor of the dropdown to accompany the role based rendering and changed the handle logout to update our authContext which will subsequently update locally saved data on the user to [null] - User data saftey!